+++ power/decimate/__init__.py
@@ -1,4 +1,5 @@
 from ._decimate import (
 	make_decimated_power,
 	m_pxy_decimated,
+	_decimate_power_obj as decimate_power_obj
 	)